Remove throwerror
When revoting in TF2, NativeVotes_IsVoteInProgress always gets skipped because of Game_IsVoteInProgress(), TF2s vote controller will stay alive a few seconds after the vote is complete. If one tries to revote right as a vote completes, it will throw an error. This could probably be worked around by adding Internal_IsVoteInProgress() as a native, but since it gets checked here anyway, we can just disable the throwerror instead.